requirements.txt
pydantic-settings
imageio
imagecodecs
cairosvg
python-magic
scrapingbee
pillow
requests
python-dotenv
pillow-heif
rich